requirements.txt
streamlit>=1.32.0
streamlit-lottie>=0.0.5
google-generativeai>=0.3.2
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv>=1.0.0
openai==1.3.0
pydantic==2.4.2
python-multipart==0.0.6
jinja2==3.1.2 